bit-pattern cast into v10->volume, and Export_BotAddSound bit-cast verbatim BotAddSound(vec3_t, int, int, int, float, float, float) export signature confirms both are genuine floats. Changed the unknown40 to attenuation (matches positioned_sound's argument order in game.h), and removed the bit-casts on both sides of the oracle and is a pure fidelity fix on the PE side (already matching).
Also restored a missing standalone a4>=0 recheck before the
a4<aasworld.d_100669BC bound (the real disasm has this as an
independent guard, separate from the earlier a4<0||a4>=numindexes
check), and applied a while(v1){if(cond)break;...} loop shape
19->17 and 20->18. sub_1001CFA0 improved 60->58 as an inlining
ripple, not directly edited.

sub_1001CE20 itself improved 59->49 insn_diffs; a further
flattened early-return form reaches 36 on ELF but regresses PE
difference confirmed not controllable via source restructuring
(4+ variants tried).

Verified: full ELF ripple check (git-stash audit_so.py diff,
zero regressions, only these 4 functions changed) and full PE

Source-commit: bde306a494b0adee48da1ec026779d5e97360273

+++ botlib/be_aas_sound.h
@@ -26,7 +26,7 @@ void sub_1001CCC0(aas_soundpool_t *a1);
 void sub_1001CD10(aas_soundpool_t *a1);
 void sub_1001CD80(aas_soundpool_t *a1);
 void __cdecl sub_1001CDD0(int a1, int a2);
-int __cdecl sub_1001CE20(float *, int, int, int, int, int, float);
+int __cdecl sub_1001CE20(float *, int, int, int, float, float, float);
 void __cdecl sub_1001CFA0(float a1);
 int __cdecl sub_1001D040(aas_soundpool_t *p);
 int __cdecl sub_1001D070(aas_soundpool_t *p);
